requirements.txt
tqdm==4.64.1
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.2
Pillow==9.5.0
requests==2.31.0
